webSynop/my_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
python-dotenv
pyodbc